Solve comments renaming files to dash-case
