+++ source/containers/main-page/context.js
@@ -1,6 +1,6 @@
 // @packages
 import React from 'react';
-import initState from './initState.json';
+import initState from './init-state.json';
 
 const homeState = React.createContext({
     state: initState,
