+++ source/resources/index.js
@@ -1,5 +1,5 @@
 // @json
-import enUs from './lang/en_US.json';
+import enUs from './lang/en-us.json';
 
 export const lang = {
     en: enUs
